split constant type defs and conversions into their own file
