+++ solver/src/core/mod.rs
@@ -1,7 +1,9 @@
+pub use cst::*;
 pub use lit::*;
 pub use var_bound::*;
 pub use variable::*;
 
+mod cst;
 mod lit;
 pub mod literals;
 pub mod state;
